refactor: extract method - signup(Member) to SignupService
